Add history files
